♻️ :: refactoring class
