--- app/src/main/java/com/example/dmsport_android/LoginRepository.kt
@@ -1,10 +1,0 @@
-package com.example.dmsport_android
-
-import retrofit2.Response
-
-class LoginRepository {
-
-//    suspend fun login(loginRequest: LoginRequest) : Response<Void> =
-//        ApiProvider.retrofit.login(loginRequest)
-
-}
