--- app/src/main/java/com/example/dmsport_android/LoginRequest.kt
@@ -1,6 +1,0 @@
-package com.example.dmsport_android
-
-data class LoginRequest(
-    val email : String,
-    val pw : String,
-)
