+++ app/src/main/java/com/example/dmsport_android/dto/request/LoginRequest.kt
@@ -1,0 +1,6 @@
+package com.example.dmsport_android.dto.request
+
+data class LoginRequest(
+    val email : String,
+    val password : String,
+)
