+++ app/src/main/java/com/example/dmsport_android/network/ApiProvider.kt
@@ -1,4 +1,4 @@
-package com.example.dmsport_android
+package com.example.dmsport_android.network
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
