+++ app/src/main/java/com/example/dmsport_android/repository/LoginRepository.kt
@@ -1,0 +1,10 @@
+package com.example.dmsport_android.repository
+
+import retrofit2.Response
+
+class LoginRepository {
+
+//    suspend fun login(loginRequest: LoginRequest) : Response<LoginResponse> =
+//        ApiProvider.retrofit.login(loginRequest)
+
+}
