+++ app/src/main/java/com/example/dmsport_android/ui/fragment/NoticeFragment.kt
@@ -1,16 +1,18 @@
-package com.example.dmsport_android
+package com.example.dmsport_android.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.dmsport_android.R
 
-class MyPageFragment : Fragment() {
-    override fun onCreateView(
+
+class NoticeFragment : Fragment() {
+      override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_my_page, container, false)
+        return inflater.inflate(R.layout.fragment_notice, container, false)
     }
 }
