+++ app/src/main/java/com/example/dmsport_android/ui/fragment/VoteFragment.kt
@@ -1,10 +1,11 @@
-package com.example.dmsport_android
+package com.example.dmsport_android.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.dmsport_android.R
 
 
 class VoteFragment : Fragment() {
